+++ src/attachments.h
@@ -32,7 +32,7 @@ int try_upload_media(struct mstdnt_storage** storage,
                      char*** media_ids);
 void cleanup_media_storages(struct session* ssn, struct mstdnt_storage* storage);
 void cleanup_media_ids(struct session* ssn, char** media_ids);
-char* construct_attachment(struct session* ssn, mstdnt_bool sensitive, struct mstdnt_attachment* att, int* str_size);
+char* construct_attachment(struct session* ssn, mstdnt_bool sensitive, struct mstdnt_attachment* att, size_t* str_size);
 char* construct_attachments(struct session* ssn, mstdnt_bool sensitive, struct mstdnt_attachment* atts, size_t atts_len, size_t* str_size);
 
 #endif // ATTACHMENTS_H
